+++ spec/spec_helper.rb
@@ -25,7 +25,7 @@ API_KEY = ENV.fetch('API_KEY', 'API_KEY_1')
 VCR.configure do |config|
   config.cassette_library_dir = "spec/fixtures/vcr_cassettes"
   config.hook_into :faraday
-  config.filter_sensitive_data("<API_KEY>") do
+  config.filter_sensitive_data('<API_KEY>') do
     API_KEY
   end
 end