+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 ### Added
 - Report Endpoint
+- Allow testing with API_KEY from ENV
 
 ## [0.0.1] - 2019-06-12
 ### Added